fix: e2e-cypress sharp not importable in browser
