+++ e2e-cypress/files/plugins/screenshot.ts
@@ -1,28 +1,7 @@
 import { createHash } from 'crypto';
 import { writeFile } from 'fs/promises';
-import sharp from 'sharp';
-
-export type ScreenshotEvidenceData = {
-  name: string;
-  path: string;
-  step: number;
-  subStep: number;
-  takenAt: string;
-};
-
-export type ScreenshotEvidenceResult = {
-  hash: string;
-  path: string;
-};
-export const isScreenshotEvidenceResult = (candidate: unknown): candidate is ScreenshotEvidenceResult =>
-  Boolean(
-    typeof candidate === 'object' &&
-      candidate &&
-      'hash' in candidate &&
-      'path' in candidate &&
-      typeof candidate.hash === 'string' &&
-      typeof candidate.path === 'string'
-  );
+import * as sharp from 'sharp';
+import { ScreenshotEvidenceData, ScreenshotEvidenceResult } from './screenshot.types';
 
 const SCREENSHOT_METADATA = {
   height: 50,
